优化验证
